refactor(core): move engine unit tests to their home modules
Relocate model-free tests that already call sizing, load_files,
token_format, and types helpers so those modules own their coverage.
engine.rs keeps Engine/pool/streaming tests and the split_pool wrapper.

Constraint: no production behavior change
Rejected: also move pool/state tests this pass | they still share engine.rs fixtures
Confidence: high
Scope-risk: narrow
