+++ crates/gigastt-core/src/inference/engine.rs
@@ -2139,156 +2139,17 @@ mod tests {
     #[cfg(feature = "diarization")]
     use crate::inference::diarization::SPEAKER_EMBEDDING_DIM;
     use crate::inference::state::aggregate_confidence;
-    use crate::inference::windows::{CHUNK_OVERLAP_SAMPLES, CHUNK_THRESHOLD_SAMPLES};
+    use crate::inference::windows::CHUNK_THRESHOLD_SAMPLES;
     use crate::inference::{
-        DEFAULT_HOTWORDS_BOOST, DecoderState, EndpointMode, EndpointReason, FeatureExtractor,
-        MAX_HOTWORD_PHRASE_CHARS, MAX_HOTWORDS_PER_REQUEST, OwnedReservation, PRED_HIDDEN, Pool,
-        PoolError, TranscribeOverrides, TranscribeResult, TranscriptAssembler, TranscriptSegment,
-        WordInfo, merge_channel_results, now_timestamp,
+        DecoderState, EndpointMode, EndpointReason, FeatureExtractor, OwnedReservation,
+        PRED_HIDDEN, Pool, PoolError, TranscriptAssembler, TranscriptSegment, WordInfo,
+        now_timestamp,
     };
     #[cfg(feature = "diarization")]
     use polyvoice::Embedder;
 
-    #[test]
-    fn test_transcribe_overrides_default_all_none() {
-        // The default overrides must be all-`None` so a request with no knobs
-        // reproduces the engine's boot behaviour byte-for-byte.
-        let o = TranscribeOverrides::default();
-        assert_eq!(o.punctuation, None);
-        assert_eq!(o.itn, None);
-        assert_eq!(o.vad, None);
-    }
-
-    #[test]
-    fn test_hotword_override_limits_constants() {
-        // Documented DoS caps used by validate_overrides and the REST 400 path.
-        assert_eq!(MAX_HOTWORDS_PER_REQUEST, 64);
-        assert_eq!(MAX_HOTWORD_PHRASE_CHARS, 64);
-        assert_eq!(DEFAULT_HOTWORDS_BOOST, 5.0);
-    }
-
-    #[test]
-    fn test_override_error_codes_stable() {
-        use crate::inference::{HotwordError, OverrideError};
-        // Stable machine-readable codes surfaced as the REST error `code`.
-        assert_eq!(OverrideError::VadNotLoaded.code(), "vad_not_loaded");
-        assert_eq!(
-            OverrideError::PunctuationNotAvailable.code(),
-            "punctuation_not_available"
-        );
-        assert_eq!(HotwordError::TooManyHotwords.code(), "too_many_hotwords");
-        assert_eq!(
-            HotwordError::PhraseTooLong.code(),
-            "hotword_phrase_too_long"
-        );
-        // Messages are non-empty and don't leak internals.
-        assert!(!OverrideError::VadNotLoaded.message().is_empty());
-        assert!(!OverrideError::PunctuationNotAvailable.message().is_empty());
-        assert!(!HotwordError::TooManyHotwords.message().is_empty());
-        assert!(!HotwordError::PhraseTooLong.message().is_empty());
-        // Display matches message().
-        assert_eq!(
-            OverrideError::VadNotLoaded.to_string(),
-            OverrideError::VadNotLoaded.message()
-        );
-        // Limit violations are client errors (400); missing models are 409.
-    }
-
-    #[test]
-    fn test_parse_cgroup_memory_limit_max_and_sentinel() {
-        use super::sizing::parse_cgroup_memory_limit;
-        assert_eq!(parse_cgroup_memory_limit("max"), None);
-        assert_eq!(parse_cgroup_memory_limit("MAX\n"), None);
-        assert_eq!(parse_cgroup_memory_limit(""), None);
-        assert_eq!(parse_cgroup_memory_limit("0"), None);
-        // v1 unlimited sentinel
-        assert_eq!(parse_cgroup_memory_limit("9223372036854771712"), None);
-        assert_eq!(
-            parse_cgroup_memory_limit("1073741824"),
-            Some(1024 * 1024 * 1024)
-        );
-        assert_eq!(
-            parse_cgroup_memory_limit("  536870912\n"),
-            Some(512 * 1024 * 1024)
-        );
-    }
-
-    #[test]
-    fn test_effective_ram_prefers_tighter_cgroup() {
-        // Pure composition: min(host, cgroup) — exercised via cap_pool_size with
-        // a 1 GiB "cgroup" budget vs larger host-class numbers.
-        let enc = 225 * 1024 * 1024;
-        let one_gib = 1024u64 * 1024 * 1024;
-        // Half of 1 GiB = 512 MiB budget; ~450 MiB/triplet → max 1 slot.
-        assert_eq!(Engine::cap_pool_size_for_ram(2, enc, one_gib), 1);
-    }
-
-    #[test]
-    fn test_cap_pool_size_for_ram_clamps_on_low_memory() {
-        // 225 MiB encoder, 2x resident => ~450 MiB/triplet. Half of 2 GiB =
-        // 1 GiB budget => floor(1024/450) = 2 slots; a request for 4 clamps.
-        let enc = 225 * 1024 * 1024;
-        let two_gib = 2 * 1024 * 1024 * 1024;
-        assert_eq!(Engine::cap_pool_size_for_ram(4, enc, two_gib), 2);
-    }
-
-    #[test]
-    fn test_cap_pool_size_for_ram_no_clamp_with_ample_ram() {
-        // 64 GiB host easily fits a pool of 4 of the same encoder.
-        let enc = 225 * 1024 * 1024;
-        let sixty_four_gib = 64u64 * 1024 * 1024 * 1024;
-        assert_eq!(Engine::cap_pool_size_for_ram(4, enc, sixty_four_gib), 4);
-    }
-
-    #[test]
-    fn test_cap_pool_size_for_ram_never_below_one() {
-        // Even a single triplet larger than the whole budget still yields 1 —
-        // the pool can't be empty; partial-load tolerance handles real OOM.
-        let huge_enc = 8 * 1024 * 1024 * 1024;
-        let small_ram = 1024 * 1024 * 1024;
-        assert_eq!(Engine::cap_pool_size_for_ram(4, huge_enc, small_ram), 1);
-    }
-
-    #[test]
-    fn test_cap_pool_size_for_ram_noop_on_unknown_inputs() {
-        // Unknown RAM or encoder size (0) => never lower the request.
-        assert_eq!(Engine::cap_pool_size_for_ram(4, 0, 8 << 30), 4);
-        assert_eq!(Engine::cap_pool_size_for_ram(4, 200 << 20, 0), 4);
-        // pool_size <= 1 is returned as-is (min 1).
-        assert_eq!(Engine::cap_pool_size_for_ram(1, 200 << 20, 1 << 30), 1);
-        assert_eq!(Engine::cap_pool_size_for_ram(0, 200 << 20, 1 << 30), 1);
-    }
-
-    #[test]
-    fn test_clamp_encoder_intra_threads() {
-        // Default request of 1 is always returned unchanged (no behavior change).
-        assert_eq!(Engine::clamp_encoder_intra_threads(2, 1, 10), 1);
-        assert_eq!(Engine::clamp_encoder_intra_threads(4, 1, 4), 1);
-
-        // Fits within budget: pool 2 x 4 threads = 8 <= 16 CPUs -> 4.
-        assert_eq!(Engine::clamp_encoder_intra_threads(2, 4, 16), 4);
-
-        // Over budget: pool 4 x 4 = 16 > 10 CPUs -> floor(10/4) = 2 per encoder.
-        assert_eq!(Engine::clamp_encoder_intra_threads(4, 4, 10), 2);
-
-        // More pooled encoders than CPUs still leaves at least 1 thread each.
-        assert_eq!(Engine::clamp_encoder_intra_threads(8, 4, 4), 1);
-
-        // Zero inputs are floored to 1 (total function, never panics/divides-by-0).
-        assert_eq!(Engine::clamp_encoder_intra_threads(0, 4, 8), 4);
-        assert_eq!(Engine::clamp_encoder_intra_threads(2, 0, 8), 1);
-        assert_eq!(Engine::clamp_encoder_intra_threads(2, 4, 0), 1);
-    }
-
-    #[test]
-    fn test_batch_split_count_clamps() {
-        use super::sizing::batch_split_count;
-        assert_eq!(batch_split_count(4, 1), 1); // typical: 1 batch, 3 stream
-        assert_eq!(batch_split_count(4, 0), 0); // split disabled
-        assert_eq!(batch_split_count(4, 10), 3); // clamped: leave 1 interactive
-        assert_eq!(batch_split_count(1, 1), 0); // can't split a single triplet
-        assert_eq!(batch_split_count(0, 1), 0); // empty pool
-        assert_eq!(batch_split_count(2, 1), 1);
+    fn word(text: &str, start: f64, end: f64) -> WordInfo {
+        WordInfo::new(text, start, end, 1.0, None)
     }
 
     #[test]
@@ -2313,434 +2174,6 @@ mod tests {
         let (pool, batch) = Engine::split_pool(vec![1u32], 1);
         assert_eq!(pool.total(), 1);
         assert!(batch.is_none());
-    }
-
-    #[test]
-    fn test_token_formatter_groups_words() {
-        // `▁` (U+2581) marks a new word; continuation tokens have no prefix.
-        let tok = Tokenizer::from_tokens(vec![
-            "\u{2581}hel".into(), // 0: new word
-            "lo".into(),          // 1: continuation
-            "\u{2581}wor".into(), // 2: new word
-            "ld".into(),          // 3: continuation
-        ]);
-        let tokens = vec![
-            decode::TokenInfo {
-                token_id: 0,
-                frame_index: 0,
-                confidence: 0.9,
-            },
-            decode::TokenInfo {
-                token_id: 1,
-                frame_index: 1,
-                confidence: 0.8,
-            },
-            decode::TokenInfo {
-                token_id: 2,
-                frame_index: 2,
-                confidence: 0.95,
-            },
-            decode::TokenInfo {
-                token_id: 3,
-                frame_index: 3,
-                confidence: 0.85,
-            },
-        ];
-        let words = TokenFormatter::tokens_to_words(&tok, &tokens, 0);
-        assert_eq!(words.len(), 2);
-        assert_eq!(words[0].word, "hello");
-        assert_eq!(words[1].word, "world");
-        // Mean confidence per word.
-        assert!((words[0].confidence - 0.85).abs() < 1e-6);
-        assert!((words[1].confidence - 0.90).abs() < 1e-6);
-        // Frame timing (SECONDS_PER_FRAME = 0.04).
-        assert!((words[0].start - 0.0).abs() < 1e-9);
-        assert!((words[0].end - 0.04).abs() < 1e-9);
-        assert!((words[1].start - 0.08).abs() < 1e-9);
-    }
-
-    #[test]
-    fn test_token_formatter_empty_tokens() {
-        let tok = Tokenizer::from_tokens(vec!["\u{2581}a".into()]);
-        assert!(TokenFormatter::tokens_to_words(&tok, &[], 0).is_empty());
-    }
-
-    #[test]
-    fn test_token_formatter_frame_offset_shifts_time() {
-        let tok = Tokenizer::from_tokens(vec!["\u{2581}x".into()]);
-        let tokens = vec![decode::TokenInfo {
-            token_id: 0,
-            frame_index: 0,
-            confidence: 1.0,
-        }];
-        let words = TokenFormatter::tokens_to_words(&tok, &tokens, 10);
-        assert_eq!(words.len(), 1);
-        // frame_offset 10 → start = 10 * 0.04 = 0.4.
-        assert!((words[0].start - 0.4).abs() < 1e-9);
-    }
-
-    fn word(text: &str, start: f64, end: f64) -> WordInfo {
-        WordInfo::new(text, start, end, 1.0, None)
-    }
-
-    fn sample_word(w: &str, start: f64, end: f64, speaker: Option<u32>) -> WordInfo {
-        WordInfo::new(w, start, end, 0.9, speaker)
-    }
-
-    #[test]
-    fn test_merge_channel_results_empty() {
-        let merged = merge_channel_results(vec![
-            TranscribeResult {
-                text: String::new(),
-                words: vec![],
-                duration_s: 0.0,
-                confidence: None,
-            },
-            TranscribeResult {
-                text: String::new(),
-                words: vec![],
-                duration_s: 0.0,
-                confidence: None,
-            },
-        ]);
-        assert!(merged.words.is_empty());
-        assert!(merged.text.is_empty());
-    }
-
-    #[test]
-    fn test_merge_channel_results_interleaved_channels() {
-        let ch0 = TranscribeResult {
-            text: String::new(),
-            words: vec![
-                sample_word("привет", 0.0, 0.4, None),
-                sample_word("как", 1.0, 1.3, None),
-            ],
-            duration_s: 1.5,
-            confidence: None,
-        };
-        let ch1 = TranscribeResult {
-            text: String::new(),
-            words: vec![sample_word("да", 0.5, 0.8, None)],
-            duration_s: 1.5,
-            confidence: None,
-        };
-        let merged = merge_channel_results(vec![ch0, ch1]);
-        assert_eq!(merged.words.len(), 3);
-        assert_eq!(merged.words[0].word, "привет");
-        assert_eq!(merged.words[0].speaker, Some(0));
-        assert_eq!(merged.words[1].word, "да");
-        assert_eq!(merged.words[1].speaker, Some(1));
-        assert_eq!(merged.words[2].word, "как");
-        assert_eq!(merged.words[2].speaker, Some(0));
-    }
-
-    #[test]
-    fn test_merge_channel_results_tie_order_by_channel() {
-        let ch0 = TranscribeResult {
-            text: String::new(),
-            words: vec![sample_word("а", 0.5, 0.7, None)],
-            duration_s: 1.0,
-            confidence: None,
-        };
-        let ch1 = TranscribeResult {
-            text: String::new(),
-            words: vec![sample_word("б", 0.5, 0.7, None)],
-            duration_s: 1.0,
-            confidence: None,
-        };
-        let merged = merge_channel_results(vec![ch0, ch1]);
-        assert_eq!(merged.words[0].word, "а");
-        assert_eq!(merged.words[0].speaker, Some(0));
-        assert_eq!(merged.words[1].word, "б");
-        assert_eq!(merged.words[1].speaker, Some(1));
-    }
-
-    #[test]
-    fn test_merge_channel_results_no_channels() {
-        let merged = merge_channel_results(vec![]);
-        assert!(merged.words.is_empty());
-        assert!(merged.text.is_empty());
-        assert_eq!(merged.duration_s, 0.0);
-    }
-
-    #[test]
-    fn test_merge_channel_results_max_duration() {
-        let ch0 = TranscribeResult {
-            text: String::new(),
-            words: vec![sample_word("a", 0.0, 0.5, None)],
-            duration_s: 5.0,
-            confidence: None,
-        };
-        let ch1 = TranscribeResult {
-            text: String::new(),
-            words: vec![sample_word("b", 0.5, 1.0, None)],
-            duration_s: 12.0,
-            confidence: None,
-        };
-        let merged = merge_channel_results(vec![ch0, ch1]);
-        assert_eq!(merged.duration_s, 12.0);
-    }
-
-    #[test]
-    fn test_stitch_first_chunk_passes_through() {
-        // An empty `merged` (the very first chunk) is returned verbatim.
-        let next = vec![word("a", 0.0, 0.5), word("b", 0.6, 1.0)];
-        let out = stitch_chunk_words(Vec::new(), next.clone(), 11.0);
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[0].word, "a");
-        assert_eq!(out[1].word, "b");
-    }
-
-    #[test]
-    fn test_stitch_dedups_overlap_no_drop_no_dup() {
-        // Two 24s windows with a 22s stride: chunk B starts at 22s, overlap
-        // [22s, 24s], seam at 23s. The word "dup" at ~22.5s is decoded by both
-        // chunks; the seam attributes it to exactly one. No unique word lost.
-        let chunk_a = vec![
-            word("first", 1.0, 1.4),    // unique to A
-            word("middle", 21.0, 21.4), // unique to A, before overlap
-            word("dup", 22.4, 22.8),    // in overlap, before seam → kept from A
-        ];
-        // B's words are already offset by its 22s start.
-        let chunk_b = vec![
-            word("dup", 22.5, 22.9),   // same word re-decoded → after-seam copy dropped
-            word("later", 25.0, 25.4), // unique to B
-            word("end", 40.0, 40.4),   // unique to B
-        ];
-        let seam_s = 22.0 + CHUNK_OVERLAP_SAMPLES as f64 / 2.0 / 16000.0; // 23.0
-        assert!((seam_s - 23.0).abs() < 1e-9);
-
-        let out = stitch_chunk_words(chunk_a, chunk_b, seam_s);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        // "dup" appears exactly once (A's copy, before the seam); nothing dropped.
-        assert_eq!(texts, vec!["first", "middle", "dup", "later", "end"]);
-        // Monotonic in `start`.
-        for w in out.windows(2) {
-            assert!(w[0].start <= w[1].start, "not monotonic: {:?}", out);
-        }
-    }
-
-    #[test]
-    fn test_stitch_drops_a_tail_past_seam() {
-        // A word decoded by A past the seam is dropped in favour of B's
-        // fuller-context copy; the back half of the overlap belongs to B.
-        let chunk_a = vec![word("keep", 22.0, 22.4), word("a_tail", 23.5, 23.9)];
-        let chunk_b = vec![word("b_seam", 23.2, 23.6), word("b_late", 30.0, 30.4)];
-        let out = stitch_chunk_words(chunk_a, chunk_b, 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(texts, vec!["keep", "b_seam", "b_late"]);
-    }
-
-    // ---- Seam invariants --------------------------------------------------
-    //
-    // `stitch_chunk_words` cuts on `start` alone: the earlier chunk keeps
-    // `start <= seam_s`, the later chunk keeps `start > seam_s`. There is no
-    // text matching across the boundary, so the *interval* of a word is
-    // irrelevant to the decision. The tests below pin what that cut does in the
-    // cases the happy-path tests above skip — including the two lossy ones.
-    // They are characterisation tests: they describe today's behaviour so a
-    // change to the seam policy shows up as a diff, not as a silent shift in
-    // long-form WER.
-    //
-    // Every word below sits within ±0.05s of the seam (or inside a 200ms gap
-    // around it, for the silence control), so nudging the seam past that margin
-    // flips at least one assertion. That is the check that these tests pin the
-    // predicate rather than passing for any cut point.
-
-    #[test]
-    fn test_stitch_straddling_word_duplicated_across_seam() {
-        // Invariant pinned: a word whose interval straddles the seam is emitted
-        // TWICE when the two chunks disagree about its start across the cut —
-        // the earlier chunk's copy starts before the seam (kept) and the later
-        // chunk's copy starts after it (also kept). The stitch has no way to
-        // notice they are the same word. This is the duplicate half of the
-        // long-form seam cost.
-        let chunk_a = vec![word("на", 22.0, 22.32), word("мосту", 22.96, 23.36)];
-        let chunk_b = vec![word("мосту", 23.04, 23.44), word("стоял", 24.0, 24.4)];
-        let out = stitch_chunk_words(chunk_a, chunk_b, 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(
-            texts,
-            vec!["на", "мосту", "мосту", "стоял"],
-            "a straddling word whose copies land on opposite sides of the seam is duplicated"
-        );
-    }
-
-    #[test]
-    fn test_stitch_straddling_word_deleted_at_seam() {
-        // Invariant pinned: the mirror-image case deletes the word outright.
-        // The earlier chunk placed its start just past the seam (dropped as
-        // "tail past the seam") while the later chunk placed it just before
-        // (dropped as "belongs to the earlier chunk"), so neither copy
-        // survives. This is the deletion half of the long-form seam cost, and
-        // it is the failure mode a text-matching stitch would have to fix.
-        let chunk_a = vec![word("на", 22.0, 22.32), word("мосту", 23.04, 23.44)];
-        let chunk_b = vec![word("мосту", 22.96, 23.36), word("стоял", 24.0, 24.4)];
-        let out = stitch_chunk_words(chunk_a, chunk_b, 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(
-            texts,
-            vec!["на", "стоял"],
-            "a straddling word can vanish entirely: dropped by both sides of the seam"
-        );
-    }
-
-    #[test]
-    fn test_stitch_word_exactly_on_seam_kept_from_earlier_chunk() {
-        // Invariant pinned: `start == seam_s` is an inclusive boundary for the
-        // earlier chunk (`<=`) and an exclusive one for the later chunk (`>`),
-        // so a word that both chunks place exactly on the seam survives once,
-        // from the earlier chunk. Confidence tags which copy won: flipping the
-        // predicate to `<` / `>=` would keep B's copy instead, same text.
-        let chunk_a = vec![WordInfo::new("шов", 23.0, 23.4, 0.5, None)];
-        let chunk_b = vec![
-            WordInfo::new("шов", 23.0, 23.4, 0.9, None),
-            word("после", 24.0, 24.4),
-        ];
-        let out = stitch_chunk_words(chunk_a, chunk_b, 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(
-            texts,
-            vec!["шов", "после"],
-            "no duplicate exactly on the seam"
-        );
-        assert_eq!(
-            out[0].confidence, 0.5,
-            "the surviving copy comes from the earlier chunk"
-        );
-    }
-
-    #[test]
-    fn test_stitch_empty_next_chunk_still_trims_tail_past_seam() {
-        // Invariant pinned: the tail trim runs unconditionally, so a chunk that
-        // decodes to nothing (silence, or a chunk the decoder emitted no tokens
-        // for) still deletes whatever the previous chunk decoded past the seam.
-        // The empty chunk is not a no-op.
-        let chunk_a = vec![word("до", 22.0, 22.4), word("хвост", 23.04, 23.44)];
-        let out = stitch_chunk_words(chunk_a, Vec::new(), 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(
-            texts,
-            vec!["до"],
-            "an empty chunk still drops the earlier chunk's post-seam tail"
-        );
-    }
-
-    #[test]
-    fn test_stitch_silence_at_seam_loses_nothing() {
-        // Lossless-gap CONTROL — read this before "fixing" it. When the seam
-        // falls inside a silent gap (no word interval touches it), the stitch
-        // must stay lossless and duplicate-free for any reasonable seam
-        // placement. That robustness is the invariant this test pins, so it is
-        // DELIBERATELY insensitive to a ±60ms seam nudge: the gap here is 200ms
-        // wide and ±60ms stays inside it, leaving the output unchanged. The
-        // tolerance is intentional — do NOT tighten the fixture to make it react
-        // to small nudges; that would delete the property being guarded.
-        //
-        // It is not vacuous: mutating the seam by +150ms moves the cut out of
-        // the gap and reddens this test (it drops "после"). The other seam tests
-        // in this block already pin the ±60ms / inclusive-boundary sensitivity;
-        // this one pins the complementary claim that silence absorbs jitter.
-        let chunk_a = vec![word("перед", 22.6, 22.9)];
-        let chunk_b = vec![word("после", 23.1, 23.5), word("конец", 24.0, 24.4)];
-        let out = stitch_chunk_words(chunk_a, chunk_b, 23.0);
-        let texts: Vec<&str> = out.iter().map(|w| w.word.as_str()).collect();
-        assert_eq!(texts, vec!["перед", "после", "конец"]);
-        for w in out.windows(2) {
-            assert!(w[0].start <= w[1].start, "not monotonic: {:?}", out);
-        }
-    }
-
-    #[test]
-    fn test_stitch_truncate_matches_retain_predicate() {
-        // `stitch_chunk_words` trims the earlier chunk with
-        // `partition_point` + `truncate` instead of `retain` (which rescanned
-        // every merged word on every chunk). On the monotonic-in-`start` lists
-        // the chunked loop actually produces, the two are the same predicate —
-        // sweep seams on and between every word boundary to show it.
-        let merged: Vec<WordInfo> = (0..40)
-            .map(|i| word(&format!("w{i}"), i as f64 * 0.5, i as f64 * 0.5 + 0.3))
-            .collect();
-        for step in 0..=80 {
-            let seam_s = step as f64 * 0.25;
-            let mut expected = merged.clone();
-            expected.retain(|w| w.start <= seam_s);
-            let got = stitch_chunk_words(merged.clone(), Vec::new(), seam_s);
-            assert_eq!(
-                got.iter().map(|w| w.word.as_str()).collect::<Vec<_>>(),
-                expected.iter().map(|w| w.word.as_str()).collect::<Vec<_>>(),
-                "diverged at seam {seam_s}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_stitch_timestamp_offset_math() {
-        // The chunked path offsets a chunk's frame indices by
-        // start_samples / (HOP_LENGTH * ENCODER_SUBSAMPLING). Verify that a word
-        // at frame 0 of a chunk starting `start_samples` in lands at the right
-        // absolute time via `tokens_to_words` (the same offset the engine feeds).
-        let tok = Tokenizer::from_tokens(vec!["\u{2581}w".into()]);
-        let tokens = vec![decode::TokenInfo {
-            token_id: 0,
-            frame_index: 0,
-            confidence: 1.0,
-        }];
-        let start_samples = 16000 * 22; // chunk starts at 22s
-        let frame_offset = start_samples / (HOP_LENGTH * ENCODER_SUBSAMPLING);
-        let words = TokenFormatter::tokens_to_words(&tok, &tokens, frame_offset);
-        assert_eq!(words.len(), 1);
-        // frame 0 + offset → absolute start == 22.0s exactly (aligned stride).
-        assert!(
-            (words[0].start - 22.0).abs() < 1e-9,
-            "got {}",
-            words[0].start
-        );
-    }
-
-    #[test]
-    fn test_finalize_pool_load_full() {
-        let r: Vec<anyhow::Result<u32>> = vec![Ok(1), Ok(2), Ok(3)];
-        assert_eq!(sizing::finalize_pool_load(r, 3, 3).unwrap(), vec![1, 2, 3]);
-    }
-
-    #[test]
-    fn test_finalize_pool_load_degraded_boots() {
-        // 2 of 4 loaded with min_size 1 → degraded pool is accepted.
-        let r: Vec<anyhow::Result<u32>> = vec![
-            Ok(1),
-            Err(anyhow::anyhow!("boom")),
-            Ok(3),
-            Err(anyhow::anyhow!("boom2")),
-        ];
-        assert_eq!(sizing::finalize_pool_load(r, 4, 1).unwrap(), vec![1, 3]);
-    }
-
-    #[test]
-    fn test_finalize_pool_load_below_min_errors() {
-        // Only 1 loaded but min_size 2 → error naming the shortfall.
-        let r: Vec<anyhow::Result<u32>> = vec![
-            Ok(1),
-            Err(anyhow::anyhow!("boom")),
-            Err(anyhow::anyhow!("boom2")),
-        ];
-        let err = sizing::finalize_pool_load(r, 3, 2).unwrap_err().to_string();
-        assert!(err.contains("loaded only 1/3"), "got: {err}");
-        assert!(err.contains("need at least 2"), "got: {err}");
-    }
-
-    #[test]
-    fn test_finalize_pool_load_all_fail_errors() {
-        let r: Vec<anyhow::Result<u32>> =
-            vec![Err(anyhow::anyhow!("a")), Err(anyhow::anyhow!("b"))];
-        assert!(sizing::finalize_pool_load(r, 2, 1).is_err());
-    }
-
-    #[test]
-    fn test_finalize_pool_load_min_clamped_to_pool() {
-        // min_size > pool_size is clamped down; a full load still succeeds.
-        let r: Vec<anyhow::Result<u32>> = vec![Ok(1), Ok(2)];
-        assert_eq!(sizing::finalize_pool_load(r, 2, 99).unwrap(), vec![1, 2]);
     }
 
     #[test]
@@ -2827,169 +2260,6 @@ mod tests {
         let dir = tempfile::tempdir().unwrap();
         let result = Engine::load_with_pool_size(dir.path().to_str().unwrap(), 1);
         assert!(matches!(result, Err(GigasttError::ModelLoad { .. })));
-    }
-
-    #[test]
-    fn test_resolve_load_variant_override_beats_disk_detection() {
-        // A directory holding BOTH the rnnt and e2e_rnnt encoders: on-disk
-        // detection returns rnnt (precedence), so without an explicit override the
-        // engine would silently ignore `--model-variant e2e_rnnt`. This is the
-        // regression this fix guards.
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join(ModelVariant::Rnnt.encoder_file()), b"x").unwrap();
-        std::fs::write(dir.path().join(ModelVariant::E2eRnnt.encoder_file()), b"x").unwrap();
-
-        // Sanity: bare on-disk detection prefers rnnt.
-        assert_eq!(
-            ModelVariant::detect_in_dir(dir.path()),
-            Some(ModelVariant::Rnnt)
-        );
-        // No override → auto-detect (rnnt precedence): behavior is unchanged.
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(None, dir.path()).unwrap(),
-            Some(ModelVariant::Rnnt)
-        );
-        // Explicit override wins over the higher-precedence head on disk.
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(
-                Some(ModelVariant::E2eRnnt),
-                dir.path()
-            )
-            .unwrap(),
-            Some(ModelVariant::E2eRnnt)
-        );
-
-        // The override is honored even when its files aren't present — the engine
-        // load then fails with a clear ModelLoad error instead of silently loading
-        // whatever else is on disk.
-        let empty = tempfile::tempdir().expect("tempdir");
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(
-                Some(ModelVariant::E2eRnnt),
-                empty.path()
-            )
-            .unwrap(),
-            Some(ModelVariant::E2eRnnt)
-        );
-        // No override + empty dir → nothing to load.
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(None, empty.path()).unwrap(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_resolve_load_variant_manifest_beats_disk_detection() {
-        // Disk has rnnt (higher precedence), but manifest selects e2e_rnnt.
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join(ModelVariant::Rnnt.encoder_file()), b"x").unwrap();
-        std::fs::write(dir.path().join(ModelVariant::E2eRnnt.encoder_file()), b"x").unwrap();
-        std::fs::write(
-            dir.path().join(crate::model::MANIFEST_FILE),
-            r#"
-architecture = "e2e_rnnt"
-[files]
-encoder = "v3_e2e_rnnt_encoder.onnx"
-decoder = "v3_e2e_rnnt_decoder.onnx"
-joint = "v3_e2e_rnnt_joint.onnx"
-vocab = "v3_e2e_rnnt_vocab.txt"
-"#,
-        )
-        .unwrap();
-
-        assert_eq!(
-            ModelVariant::detect_in_dir(dir.path()),
-            Some(ModelVariant::Rnnt)
-        );
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(None, dir.path()).unwrap(),
-            Some(ModelVariant::E2eRnnt)
-        );
-        // CLI override still wins over the manifest architecture.
-        assert_eq!(
-            crate::inference::load_files::resolve_load_variant(
-                Some(ModelVariant::Rnnt),
-                dir.path()
-            )
-            .unwrap(),
-            Some(ModelVariant::Rnnt)
-        );
-    }
-
-    #[test]
-    fn test_resolve_load_variant_invalid_manifest_is_error() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(
-            dir.path().join(crate::model::MANIFEST_FILE),
-            "not = [valid\n",
-        )
-        .unwrap();
-        assert!(crate::inference::load_files::resolve_load_variant(None, dir.path()).is_err());
-        // Even with an override, a corrupt manifest is a hard error.
-        assert!(
-            crate::inference::load_files::resolve_load_variant(
-                Some(ModelVariant::Rnnt),
-                dir.path()
-            )
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn test_resolved_model_files_without_manifest_match_variant_basenames() {
-        // Regression: no manifest.toml → byte-identical hardcoded names.
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("v3_rnnt_encoder.onnx"), b"fp32").unwrap();
-        std::fs::write(dir.path().join("v3_rnnt_encoder_int8.onnx"), b"int8").unwrap();
-
-        let files = ResolvedModelFiles::resolve(dir.path(), ModelVariant::Rnnt).unwrap();
-        assert_eq!(
-            files.encoder.file_name().unwrap(),
-            "v3_rnnt_encoder_int8.onnx"
-        );
-        assert!(files.using_int8);
-        assert_eq!(
-            files.decoder.as_ref().unwrap().file_name().unwrap(),
-            "v3_rnnt_decoder.onnx"
-        );
-        assert_eq!(
-            files.joint.as_ref().unwrap().file_name().unwrap(),
-            "v3_rnnt_joint.onnx"
-        );
-        assert_eq!(files.vocab.file_name().unwrap(), "v3_vocab.txt");
-    }
-
-    #[test]
-    fn test_resolved_model_files_manifest_encoder_prefers_int8() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("custom_enc.onnx"), b"fp32").unwrap();
-        std::fs::write(dir.path().join("custom_enc_int8.onnx"), b"int8").unwrap();
-        std::fs::write(
-            dir.path().join(crate::model::MANIFEST_FILE),
-            r#"
-architecture = "rnnt"
-[files]
-encoder = "custom_enc.onnx"
-encoder_int8 = "custom_enc_int8.onnx"
-decoder = "custom_dec.onnx"
-joint = "custom_joint.onnx"
-vocab = "custom_vocab.txt"
-"#,
-        )
-        .unwrap();
-
-        let files = ResolvedModelFiles::resolve(dir.path(), ModelVariant::Rnnt).unwrap();
-        assert_eq!(
-            files.encoder.file_name().unwrap(),
-            "custom_enc_int8.onnx",
-            "manifest INT8 encoder must win when present on disk"
-        );
-        assert!(files.using_int8);
-        assert_eq!(
-            files.decoder.as_ref().unwrap().file_name().unwrap(),
-            "custom_dec.onnx"
-        );
-        assert_eq!(files.vocab.file_name().unwrap(), "custom_vocab.txt");
     }
 
     // ---- Pool tests (B.7) ---------------------------------------------------
@@ -3533,145 +2803,6 @@ vocab = "custom_vocab.txt"
     }
 
     #[test]
-    fn test_probe_or_rebuild_keeps_state_when_probe_passes() {
-        let rebuilt = std::cell::Cell::new(false);
-        let result = sizing::probe_or_rebuild(
-            7u32,
-            |v| {
-                assert_eq!(*v, 7);
-                Ok(())
-            },
-            |_, _| {
-                rebuilt.set(true);
-                Ok(99)
-            },
-        )
-        .expect("healthy state must survive unchanged");
-        assert_eq!(result, 7);
-        assert!(!rebuilt.get(), "rebuild must not run when the probe passes");
-    }
-
-    #[test]
-    fn test_probe_or_rebuild_rebuilds_when_probe_fails() {
-        let result = sizing::probe_or_rebuild(
-            1u32,
-            |v| {
-                if *v == 1 {
-                    Err(anyhow::anyhow!("first probe failed"))
-                } else {
-                    Ok(())
-                }
-            },
-            |old, probe_err| {
-                assert_eq!(old, 1, "rebuild receives the failed state");
-                assert!(
-                    probe_err.to_string().contains("first probe failed"),
-                    "rebuild receives the probe error for logging"
-                );
-                Ok(2)
-            },
-        )
-        .expect("rebuilt state passing the probe is OK");
-        assert_eq!(result, 2);
-    }
-
-    #[test]
-    fn test_probe_or_rebuild_propagates_rebuild_error() {
-        let result = sizing::probe_or_rebuild(
-            1u32,
-            |_| Err(anyhow::anyhow!("probe failed")),
-            |_, _| Err(anyhow::anyhow!("rebuild failed")),
-        );
-        let err = result.expect_err("rebuild failure must be fatal");
-        assert!(err.to_string().contains("rebuild failed"));
-    }
-
-    #[test]
-    fn test_probe_or_rebuild_fails_when_rebuilt_state_fails_probe() {
-        let result = sizing::probe_or_rebuild(
-            1u32,
-            |_| Err(anyhow::anyhow!("always fails")),
-            |_, _| Ok(2u32),
-        );
-        assert!(
-            result.is_err(),
-            "a rebuilt state that still fails the probe must be a hard error"
-        );
-    }
-
-    #[test]
-    fn test_encoder_model_path_prefers_int8_when_present() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("v3_e2e_rnnt_encoder.onnx"), b"fp32").unwrap();
-        std::fs::write(dir.path().join("v3_e2e_rnnt_encoder_int8.onnx"), b"int8").unwrap();
-        let path = encoder_model_path(dir.path(), ModelVariant::E2eRnnt);
-        assert_eq!(
-            path.file_name().unwrap(),
-            "v3_e2e_rnnt_encoder_int8.onnx",
-            "INT8 encoder must win when both files exist"
-        );
-    }
-
-    #[test]
-    fn test_encoder_model_path_fp32_only_is_rejected() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("v3_e2e_rnnt_encoder.onnx"), b"fp32").unwrap();
-        // Without INT8, resolve fails; helper still reports the INT8 basename
-        // (never the FP32 file).
-        let path = encoder_model_path(dir.path(), ModelVariant::E2eRnnt);
-        assert_eq!(path.file_name().unwrap(), "v3_e2e_rnnt_encoder_int8.onnx");
-        assert!(
-            ResolvedModelFiles::from_variant(dir.path(), ModelVariant::E2eRnnt).is_err(),
-            "FP32-only install must not be loadable"
-        );
-    }
-
-    #[test]
-    fn test_encoder_model_path_rnnt_prefers_int8() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("v3_rnnt_encoder.onnx"), b"fp32").unwrap();
-        std::fs::write(dir.path().join("v3_rnnt_encoder_int8.onnx"), b"int8").unwrap();
-        let path = encoder_model_path(dir.path(), ModelVariant::Rnnt);
-        assert_eq!(
-            path.file_name().unwrap(),
-            "v3_rnnt_encoder_int8.onnx",
-            "INT8 rnnt encoder must win when both files exist"
-        );
-    }
-
-    #[test]
-    fn test_encoder_model_path_rnnt_fp32_only_is_rejected() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("v3_rnnt_encoder.onnx"), b"fp32").unwrap();
-        assert!(
-            ResolvedModelFiles::from_variant(dir.path(), ModelVariant::Rnnt).is_err(),
-            "FP32-only install must not be loadable"
-        );
-    }
-
-    #[test]
-    fn test_encoder_model_path_uses_manifest_int8_when_present() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pack_enc.onnx"), b"fp32").unwrap();
-        std::fs::write(dir.path().join("pack_enc_int8.onnx"), b"int8").unwrap();
-        std::fs::write(
-            dir.path().join(crate::model::MANIFEST_FILE),
-            r#"
-architecture = "rnnt"
-[files]
-encoder = "pack_enc.onnx"
-encoder_int8 = "pack_enc_int8.onnx"
-decoder = "pack_dec.onnx"
-joint = "pack_joint.onnx"
-vocab = "pack_vocab.txt"
-"#,
-        )
-        .unwrap();
-        let path = encoder_model_path(dir.path(), ModelVariant::Rnnt);
-        assert_eq!(path.file_name().unwrap(), "pack_enc_int8.onnx");
-    }
-
-    #[test]
     fn test_pool_sequential_checkouts_visit_every_item() {
         // Engine::warmup relies on this FIFO property: `total()` sequential
         // checkout/checkin cycles touch every pooled item exactly once.
@@ -3704,23 +2835,6 @@ vocab = "pack_vocab.txt"
     // ---- More pure (no-model) coverage -------------------------------------
 
     #[test]
-    fn test_finalize_pool_load_degraded_includes_error_detail() {
-        // The degraded-pool branch logs the first error; exercise the
-        // `first_err` formatting path (the loaded triplets are still returned).
-        let r: Vec<anyhow::Result<u32>> = vec![Ok(1), Err(anyhow::anyhow!("first failure cause"))];
-        assert_eq!(sizing::finalize_pool_load(r, 2, 1).unwrap(), vec![1]);
-    }
-
-    #[test]
-    fn test_finalize_pool_load_below_min_no_errors_when_all_ok_but_short() {
-        // No Err entries, but fewer results than pool_size with min above the
-        // loaded count → still errors (loaded count is what matters).
-        let r: Vec<anyhow::Result<u32>> = vec![Ok(1)];
-        let err = sizing::finalize_pool_load(r, 3, 2).unwrap_err().to_string();
-        assert!(err.contains("loaded only 1/3"), "got: {err}");
-    }
-
-    #[test]
     fn test_transcript_assembler_set_words_overwrites() {
         // The sliding-window streaming path overwrites (not appends) on each
         // re-decode via `set_words`. A second call replaces the first.
@@ -3743,38 +2857,6 @@ vocab = "pack_vocab.txt"
         assert!(!asm.is_empty());
         asm.set_words(vec![]);
         assert!(asm.is_empty(), "empty set_words clears the accumulation");
-    }
-
-    #[test]
-    fn test_token_formatter_last_word_empty_confidences_defaults_to_one() {
-        // A word whose only token is a bare boundary marker (`▁`, no body)
-        // contributes no confidence sample; a following real word that itself
-        // has no recorded confidences must default to 1.0 on the final-emit
-        // path. We build a vocab whose tokens are pure boundary markers so the
-        // `clean` body is empty and no confidence is pushed.
-        let tok = Tokenizer::from_tokens(vec![
-            "\u{2581}real".into(), // 0: a real word
-            "\u{2581}".into(),     // 1: bare boundary, empty body
-        ]);
-        let tokens = vec![
-            decode::TokenInfo {
-                token_id: 0,
-                frame_index: 0,
-                confidence: 0.7,
-            },
-            // A bare boundary token forces emission of "real" (mid-loop emit),
-            // then contributes nothing to a new word.
-            decode::TokenInfo {
-                token_id: 1,
-                frame_index: 1,
-                confidence: 0.5,
-            },
-        ];
-        let words = TokenFormatter::tokens_to_words(&tok, &tokens, 0);
-        // Only "real" is emitted; the trailing bare boundary leaves no word.
-        assert_eq!(words.len(), 1);
-        assert_eq!(words[0].word, "real");
-        assert!((words[0].confidence - 0.7).abs() < 1e-6);
     }
 
     #[test]
